feat: Derive identifier types from initializers for computed access
Add deriveTypeFromInitializer to handle variables without explicit
type annotations (e.g., `const arr = createArray()`). This fixes
computed member access classification where object.inferredType was
undefined.

Changes:
- Add deriveCallReturnType: extracts return type from call declarations
- Add deriveNewExpressionType: extracts type from new expressions
- Add deriveTypeFromInitializer: handles calls, new, identifiers, literals
- Enhance deriveIdentifierType to use initializer type as fallback
- Update golden files for deterministic typing output

E2E tests: 19 failing → 8 failing (11 more tests now passing)

+++ packages/emitter/testcases/common/expected/classes/field-inference/Counter.cs
@@ -2,7 +2,7 @@ namespace TestCases.common.classes.fieldinference
 {
     public class Counter
     {
-        public double count = 0;
+        public int count = 0;
 
         public string name = "default";
